fix: mirror env vars as sbt setting for caching
